chore: reorganize pages and components into package folders
