+++ src/components/DeviceTable/DeviceTable.module.css
