+++ tests/DeviceTable.test.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { render, screen } from '@testing-library/react';
 import '@testing-library/jest-dom';
 import DeviceTable from '../src/components/DeviceTable';
-import styles from '../src/components/DeviceTable.module.css';
+import styles from '../src/components/DeviceTable/DeviceTable.module.css';
 
 const devices = {
   dev1: {
